I turned off multi-threading for the moment, because it's not being caused by the timer, and I'm not sure why it won't play nice with the view.
But I re-wrote the buildSomething function. We now
have no choice but to simualate turn by turn, except
in the case of moving the robber, when it is not
computationally feasible for us to simulate an
entire turn at once.

Roads are still broken.

I guess this commit can be seen as an improvement. ~Chris